+++ contracts/socket/andromeda-socket-astroport/src/astroport.rs
@@ -22,6 +22,10 @@ use andromeda_socket::astroport::{
 
 pub const ASTROPORT_MSG_SWAP_ID: u64 = 1;
 pub const ASTROPORT_MSG_FORWARD_ID: u64 = 2;
+pub const ASTROPORT_MSG_CREATE_PAIR_ID: u64 = 3;
+pub const ASTROPORT_MSG_CREATE_PAIR_AND_PROVIDE_LIQUIDITY_ID: u64 = 4;
+pub const ASTROPORT_MSG_PROVIDE_LIQUIDITY_ID: u64 = 5;
+pub const ASTROPORT_MSG_WITHDRAW_LIQUIDITY_ID: u64 = 6;
 
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn execute_swap_astroport_msg(